[PlaybackSerialiser] Renames PlaybackSerialiser2 to PlaybackSerialiser
- Renamed PlaybackSerialiser to PlaybackSerialiserOld
- Added documentation to PlaybackSerialiser
- Removed BetaFlavorTest as that is tested in SerialiserFlavorBase
- Fixed tests failing
- Organised some imports
